backend/requirements.txt
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
fastapi
uvicorn
python-jose==3.3.0
passlib==1.7.4
ecdsa==0.19.0
rsa==4.9
setuptools>=65.5.1
python-dotenv
chromadb
pdfplumber
beautifulsoup4
requests

